+++ example/ios/InstabugTests/BugReportingApiTests.m
@@ -174,5 +174,29 @@
 
     OCMVerify([self.mBugReporting setCommentMinimumCharacterCountForReportTypes:IBGBugReportingReportTypeBug | IBGBugReportingReportTypeFeedback | IBGBugReportingReportTypeQuestion withLimit:limit.intValue]);
 }
+- (void) testSetProactiveReportingConfigurations {
+  BOOL enabled = true;
+  NSNumber* gap = @2;
+  NSNumber* delay = @4;
+    NSNumber *enabledNum = [NSNumber numberWithBool:enabled];
+    FlutterError *error;
+
+  IBGProactiveReportingConfigurations *configurations = [[IBGProactiveReportingConfigurations alloc] init];
+  configurations.enabled = enabled; //Enable/disable
+  configurations.gapBetweenModals = gap; // Time in seconds
+  configurations.modalDelayAfterDetection = delay; // Time in seconds
+  
+  OCMStub([self.mBugReporting setProactiveReportingConfigurations:OCMOCK_ANY]);
+
+    [self.api setProactiveReportingConfigurationsEnabled:enabledNum gapBetweenModals:gap modalDelayAfterDetection:delay error:&error];
+  
+  // Verify that the method is called with the correct properties (using OCMArg to match properties)
+  OCMVerify([self.mBugReporting setProactiveReportingConfigurations:[OCMArg checkWithBlock:^BOOL(id obj) {
+      IBGProactiveReportingConfigurations *config = (IBGProactiveReportingConfigurations *)obj;
+      return config.enabled == enabled &&
+             [config.gapBetweenModals isEqualToNumber:gap] &&
+             [config.modalDelayAfterDetection isEqualToNumber:delay];
+    }]]);
+}
 
 @end