+++ test/bug_reporting_test.dart
@@ -2,6 +2,7 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:instabug_flutter/instabug_flutter.dart';
 import 'package:instabug_flutter/src/generated/bug_reporting.api.g.dart';
+import 'package:instabug_flutter/src/models/proactive_reporting_config.dart';
 import 'package:instabug_flutter/src/utils/enum_converter.dart';
 import 'package:instabug_flutter/src/utils/ibg_build_info.dart';
 import 'package:mockito/annotations.dart';
@@ -99,7 +100,7 @@ void main() {
 
   test(
     '[setVideoRecordingFloatingButtonPosition] should call host method',
-    () async {
+        () async {
       const position = Position.topLeft;
 
       await BugReporting.setVideoRecordingFloatingButtonPosition(position);
@@ -199,4 +200,19 @@ void main() {
       mHost.setCommentMinimumCharacterCount(count, reportTypes.mapToString()),
     ).called(1);
   });
+
+  test('[setProactiveReportingConfigurations] should call host method',
+          () async {
+        await BugReporting.setProactiveReportingConfigurations(
+            ProactiveReportingConfigsBuilder()
+                .setGapBetweenModals(1)
+                .setModalDelayAfterDetection(1)
+                .isEnabled(true)
+                .build());
+
+        verify(
+            mHost.setProactiveReportingConfigurations(
+                true, 1, 1
+            )).called(1);
+        });
 }